fix: store ratings.note and rating_seasons.rating as NUMERIC(3,1)
A note is a one-decimal value by contract, but both columns were REAL
(float32), which cannot represent one decimal place exactly: 5.6 stored as
5.5999999046325684, and reading it back widened that into the full binary
expansion. Migration 006 repaired the *values* this produced; this finishes
the job with migration 009, changing the *type* so the class of bug cannot
reopen no matter what gets written to it. NUMERIC(3,1) rounds any residual
float32 noise as it converts, for the same reason 006's own repair was safe:
the target value is a total function of the stored value alone.

The whole note path is float64 now, with no pgtype.Numeric anywhere: a sqlc
override maps numeric to float64 directly, the same way DOUBLE PRECISION
already maps to float64 for rating_aggregate, and pgx/v5's default type map
wraps a bare float64 for the numeric codec natively. A rating's JSON always
renders its note as a clean decimal instead of a widened float or a
{"Int":...,"Exp":...} struct.

internal/activity's noteValue helper (added in #34 to paper over the same
float32 widening at the payload level) is removed along with the four
constructors' need for it: nothing upstream of them is imprecise anymore.

Covered by: an exact round-trip through the store for values inexact as
float32; a migration-level test seeding noisy REAL rows before 009 runs,
including that `note = 5.6` matches nothing pre-migration (REAL vs. a numeric
literal promotes to double precision) and matches exactly post-migration; a
raw-body JSON assertion on the API response; and a DB-independent unit test
for the TV-series mean rounding, added because the integration-level version
of that test stopped being able to tell a working rounding function apart
from a broken one once NUMERIC(3,1) started rounding on write regardless.

+++ internal/database/ratings.sql.go
@@ -234,7 +234,7 @@ type InsertRatingParams struct {
 	TitleID   string
 	UserID    string
 	GroupID   string
-	Note      float32
+	Note      float64
 	CreatedAt pgtype.Timestamptz
 	UpdatedAt pgtype.Timestamptz
 }
@@ -273,7 +273,7 @@ INSERT INTO rating_seasons (
 type InsertRatingSeasonParams struct {
 	RatingID  string
 	Season    string
-	Rating    float32
+	Rating    float64
 	AddedAt   pgtype.Timestamptz
 	UpdatedAt pgtype.Timestamptz
 }
@@ -299,7 +299,7 @@ RETURNING id, title_id, user_id, note, created_at, updated_at, group_id
 type UpdateRatingRowParams struct {
 	ID        string
 	UserID    string
-	Note      float32
+	Note      float64
 	UpdatedAt pgtype.Timestamptz
 }
